Add require_pragma option
This patch adds a configuration option for requiring files to have a
pragma comment at the beginning of the file in order to reformat the
file. This behaviour is useful for a slow adoption of the tool instead
of requiring the whole codebase to be converted at once.

The patch also contains tests for the different kinds of supported
pragma comments, based on the profile defined for the project.