fix(formatter): Avoid rewriting the file if it has not been changed
We use djLint as part of treefmt [1] to reformat HTML templates
tree-wide, and treefmt looks at file modification time to determine
if it was reformatted or not. And with --fail-on-change option
passed, it exits with status 1 when it thinks files are modified.

This patch makes djLint not rewrite the file if there are no changes
to write back, which makes treefmt --fail-on-change do the right
thing.

[1]: https://github.com/numtide/treefmt