perf: Cache compiled regexes
Even though the regex module has a cache, it's access is not that fast.
E.g. re.sub is a combination of re._compile and pattern.sub. re._compile
is checking the flags for e.g. DEBUG values or the verbosity. It uses
an enum for the flags. enum.__and__ is quite slow, so even for a cache hit
we have around two to three enum.__and__ calls. This patch caches commonly used
regexes. The naming probably has to be adjusted.

Commonly used is defined as: pattern._compile is hit with this pattern+flags
combination more than 1000 times in the netbox repo. That somewhat balances out
the time needed for compilation, the runtime speed and the maintenance effort.