docs: Avoid raw `<br>` tag in H036 description
The literal `<br>` in the Markdown documentation was being rendered as
HTML, rather than as text, making the description of rule H036 on the
website confusing.

Instead, use backticks to make it a code block, so as to be consistent
with other rules that describe tags but do not wrap them in angle
brackets.

Commit cefa8401708d9c6b32675b268111ccd1e9991865
Issue #571 [FEATURE] No BR
PR #588 feat(linter): added rule H036. Avoid use of <br> tags. Disabled by default